# docs/swagger.yaml
basePath: /
definitions:
  response.BaseResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  response.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: integer
    type: object
  response.ResponseBodyEvent:
    properties:
      description:
        type: string
      imgUrl:
        type: string
      name:
        type: string
      viewed:
        type: integer
    type: object
  response.ResponseBodyEventList:
    properties:
      events:
        items:
          $ref: '#/definitions/response.ResponseBodyEvent'
        type: array
    type: object
  response.ResponseBodyUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      surname:
        type: string
    type: object
host: bmstusasa.herokuapp.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: TP_2021_GO TEAM YO
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BMSTUSA API
  version: "1.0"
paths:
  /events:
    get:
      description: '"Список мероприятий"'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseBodyEventList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      tags:
      - Events
  /signin:
    post:
      consumes:
      - application/json
      description: '"Авторизация"'
      parameters:
      - description: Account Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/response.ResponseBodyUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                body:
                  $ref: '#/definitions/response.ResponseBodyUser'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Регистрация
      parameters:
      - description: Account Info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/response.ResponseBodyUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                body:
                  $ref: '#/definitions/response.ResponseBodyUser'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      tags:
      - auth
  /user:
    get:
      description: '"Главная страница"'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.BaseResponse'
      tags:
      - auth
schemes:
- https
swagger: "2.0"